+++ x11rb-protocol/src/x11_utils.rs
@@ -507,3 +507,45 @@ impl TryIntoUSize for i8 {}
 impl TryIntoUSize for i16 {}
 impl TryIntoUSize for i32 {}
 impl TryIntoUSize for i64 {}
+
+/// Has the BigRequests extension been enabled?
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum BigRequests {
+    /// The BigRequests extension has been enabled.
+    Enabled,
+    /// The BigRequests extension has not been enabled.
+    NotEnabled,
+}
+
+/// Parse the given input for a RequestHeader and the remaining input.
+pub fn parse_request_header(
+    input: &[u8],
+    big_requests_enabled: BigRequests,
+) -> Result<(RequestHeader, &[u8]), ParseError> {
+    let (major_opcode, remaining) = u8::try_parse(input)?;
+    let (minor_opcode, remaining) = u8::try_parse(remaining)?;
+    let (length, remaining) = u16::try_parse(remaining)?;
+    let (remaining_length, finally_remaining) = if length == 0 {
+        if big_requests_enabled == BigRequests::NotEnabled {
+            return Err(ParseError::InvalidValue);
+        }
+
+        let (length, remaining) = u32::try_parse(remaining)?;
+        if length < 2 {
+            return Err(ParseError::InvalidValue);
+        }
+        // Adjust length for the size of this header (two 4 byte units).
+        (length - 2, remaining)
+    } else {
+        // Adjust length for the size of this header (one 4 byte unit).
+        (u32::from(length) - 1, remaining)
+    };
+    Ok((
+        RequestHeader {
+            major_opcode,
+            minor_opcode,
+            remaining_length,
+        },
+        finally_remaining,
+    ))
+}
